refactor indexer creation
